security_manager.py
import google.auth
from google.cloud import Security

class SecurityManager:
    def __init__(self):
        self.credentials, self.project = google.auth.default()

    # Add methods to interact with the respective GCP service here
