requirements.txt
scrapy==2.5.0
beautifulsoup4==4.9.3
selenium==3.141.0
pandas==1.2.4
pytest==6.2.4
google-cloud-bigquery==2.20.0
pymongo==3.11.4